staging update

+++ src/modules/proxyMember.ts
@@ -1,29 +1,43 @@
-import DbTx from '@modules/dbTx'
 import { Models } from '@dbModels'
 import { type HexAddress, IMetricAction, type NetworksEnum } from '@types'
 import Web3Helper from '@helpers/web3'
 import logger from '@logger'
 import EnsHelper from '@helpers/ens'
 import type Member from '@models/schema/member'
-import DbOperations from '@models/utils/dbOperations'
+import type MemberBalance from '@models/schema/memberBalance'
+import type MemberMetrics from '@models/schema/memberMetrics'
+import DbTx from '@modules/dbTx'
 
 const llo = logger.logMeta.bind(null, { service: 'modules:ProxyMember' })
 
 export const ProxyMember = {
-  createMember: async (memberAddress: HexAddress): Promise<Member> => {
+  createMember: async (memberAddress: HexAddress): Promise<Member | null> => {
     const parsedMemberAddress = Web3Helper.parseAddress(memberAddress) || memberAddress
-    const existingMember = await Models.Member.findExistingLog({ address: parsedMemberAddress })
+    if (!parsedMemberAddress) return null
 
-    if (!existingMember) {
-      const rawMember = {
-        address: memberAddress,
-        ens: await EnsHelper.getEnsWithUniversalResolver(memberAddress),
-      }
+    try {
+      return await DbTx.executeTxFn(async ({ session }) => {
+        const existingMember = await Models.Member.findExistingLog({ address: parsedMemberAddress }, { session })
 
-      return await DbOperations.createDocument(Models.Member, rawMember, rawMember, 'New Member', llo)
+        if (!existingMember) {
+          const rawMember = {
+            address: memberAddress,
+            ens: await EnsHelper.getEnsWithUniversalResolver(memberAddress),
+          }
+
+          const newMember = await Models.Member.create(rawMember, { session })
+          await session.commitTransaction()
+          await session.endSession()
+          logger.verbose('Create document - New Member', llo({ documentId: newMember.id }))
+          return newMember
+        }
+
+        return existingMember
+      })
+    } catch (error) {
+      logger.error('Error creating new member', llo({ error, memberAddress }))
+      return null
     }
-
-    return existingMember
   },
 
   createMetrics: async ({
@@ -34,15 +48,27 @@ export const ProxyMember = {
     address: HexAddress
     pluginAddress: HexAddress
     network: NetworksEnum
-  }) => {
-    const metrics = await Models.MemberMetrics.findOne({ address, pluginAddress, network })
+  }): Promise<MemberMetrics | null> => {
+    try {
+      return await DbTx.executeTxFn(async ({ session }) => {
+        const metrics = await Models.MemberMetrics.findOne({ address, pluginAddress, network }, null, { session })
 
-    if (!metrics) {
-      const data = { address, pluginAddress, network }
-      return await DbOperations.createDocument(Models.MemberMetrics, data, data, 'New Member metrics', llo)
+        if (metrics) {
+          return metrics
+        }
+
+        const rawMetrics = { address, pluginAddress, network }
+
+        const newMetrics = await Models.MemberMetrics.create(rawMetrics, { session })
+        await session.commitTransaction()
+        await session.endSession()
+        logger.verbose('Create document - New MemberMetrics', llo({ documentId: newMetrics.id }))
+        return newMetrics
+      })
+    } catch (error) {
+      logger.error('Error creating new member metrics', llo({ error, address, pluginAddress }))
+      return null
     }
-
-    return metrics
   },
 
   getBalances: async ({
@@ -53,15 +79,26 @@ export const ProxyMember = {
     address: HexAddress
     tokenAddress: HexAddress
     network: NetworksEnum
-  }) => {
-    const token = await Models.MemberBalance.findByAddressAndToken({ address, tokenAddress, network })
+  }): Promise<MemberBalance | null> => {
+    try {
+      return await DbTx.executeTxFn(async ({ session }) => {
+        const token = await Models.MemberBalance.findByAddressAndToken({ address, tokenAddress, network }, { session })
 
-    if (!token) {
-      const data = { address, tokenAddress, network }
-      return await DbOperations.createDocument(Models.MemberBalance, data, data, 'New Member balance', llo)
+        if (token) {
+          return token
+        }
+
+        const data = { address, tokenAddress, network }
+        const memberBalance = await Models.MemberBalance.create(data, { session })
+        await session.commitTransaction()
+        await session.endSession()
+        logger.verbose('Create document - New MemberBalance', llo({ documentId: memberBalance.id }))
+        return memberBalance
+      })
+    } catch (error) {
+      logger.error('Error getting member balances', llo({ error, address, tokenAddress }))
+      return null
     }
-
-    return token
   },
 
   updateActivity: async ({
@@ -74,32 +111,39 @@ export const ProxyMember = {
     pluginAddress: HexAddress
     blockNumber: number
     network: NetworksEnum
-  }): Promise<Member> => {
-    const [member, memberMetrics, blockTimestamp] = await Promise.all([
-      ProxyMember.createMember(memberAddress),
-      ProxyMember.createMetrics({
-        address: memberAddress,
-        pluginAddress,
-        network,
-      }),
-      Web3Helper.getBlockTimestamp(blockNumber, network),
-    ])
+  }): Promise<Member | null> => {
+    try {
+      const [member, memberMetrics, blockTimestamp] = await Promise.all([
+        ProxyMember.createMember(memberAddress),
+        ProxyMember.createMetrics({
+          address: memberAddress,
+          pluginAddress,
+          network,
+        }),
+        Web3Helper.getBlockTimestamp(blockNumber, network),
+      ])
 
-    const updateFields: Partial<Member> = {
-      lastActivity: blockTimestamp,
+      if (!member || !memberMetrics) return null
+
+      const updateFields: Partial<Member> = {
+        lastActivity: blockTimestamp,
+      }
+
+      if (!member?.firstActivity) {
+        updateFields.firstActivity = blockTimestamp
+      }
+
+      return await DbTx.executeTxFn(async ({ session }) => {
+        const updatedMember = await memberMetrics?.update(updateFields, { session })
+        await session.commitTransaction()
+        await session.endSession()
+        logger.verbose('Update Member activity', llo({ documentId: updatedMember.id }))
+        return updatedMember
+      })
+    } catch (error) {
+      logger.error('Error updating member activity', llo({ error, memberAddress, pluginAddress, blockNumber, network }))
+      return null
     }
-
-    if (!member.firstActivity) {
-      updateFields.firstActivity = blockTimestamp
-    }
-
-    return await DbOperations.updateDocument(
-      memberMetrics,
-      updateFields,
-      { logId: member.id },
-      'Update Member activity',
-      llo,
-    )
   },
 
   updateMetricsByAction: async (
@@ -113,12 +157,14 @@ export const ProxyMember = {
       pluginAddress: HexAddress
       network: NetworksEnum
     },
-  ): Promise<Member> => {
+  ) => {
     const metrics = await ProxyMember.createMetrics({
       address: memberAddress,
       pluginAddress,
       network,
     })
+
+    if (!metrics) return
 
     const metricActionsMap = {
       [IMetricAction.increaseProposalCount]: metrics.increaseProposalCount,
@@ -130,17 +176,15 @@ export const ProxyMember = {
     const metricUpdateFn = metricActionsMap[metricAction]
 
     if (metricUpdateFn) {
-      const updatedMetrics = await DbTx.executeTxFn(async ({ session }) => {
+      await DbTx.executeTxFn(async ({ session }) => {
         const logDb = await metricUpdateFn.call(metrics, 1, { session })
         await session.commitTransaction()
+        await session.endSession()
         logger.verbose('Updated Member DAO metrics', { logId: logDb.id })
-        return logDb
       })
-
-      return updatedMetrics
+    } else {
+      logger.error('Unsupported metric action', llo({ metricAction, memberAddress, pluginAddress, network }))
     }
-
-    return metrics
   },
 
   addToDao: async (params: {
@@ -148,30 +192,37 @@ export const ProxyMember = {
     daoAddress: HexAddress
     pluginAddress: HexAddress
     network: NetworksEnum
-  }): Promise<Member> => {
+  }): Promise<Member | null> => {
+    const memberAddress = Web3Helper.parseAddress(params.memberAddress)
+    if (!memberAddress) return null
+
     const member = await ProxyMember.createMember(params.memberAddress)
-
-    const queryParams = {
-      memberAddress: member.address,
-      daoAddress: params.daoAddress,
-      pluginAddress: params.pluginAddress,
-      network: params.network,
-    }
-    const existingDao = await Models.DaoMemberMapping.findMapping(queryParams)
-
-    if (existingDao) {
-      return member
+    if (!member) {
+      logger.error('Failed to add member to dao', llo({ params }))
+      return null
     }
 
-    await DbOperations.createDocument(
-      Models.DaoMemberMapping,
-      queryParams,
-      { logId: member.id },
-      'New DaoMemberMapping',
-      llo,
-    )
+    try {
+      return await DbTx.executeTxFn(async ({ session }) => {
+        const queryParams = {
+          memberAddress: params.memberAddress,
+          daoAddress: params.daoAddress,
+          pluginAddress: params.pluginAddress,
+          network: params.network,
+        }
+        const existingDao = await Models.DaoMemberMapping.findMapping(queryParams, { session })
 
-    return member
+        if (!existingDao) {
+          await Models.DaoMemberMapping.create(queryParams, { session })
+          await session.commitTransaction()
+          await session.endSession()
+        }
+        return member
+      })
+    } catch (error) {
+      logger.error('Error in addToDao', llo({ error, params }))
+      return null
+    }
   },
 
   removeFromDao: async (params: {
@@ -179,25 +230,34 @@ export const ProxyMember = {
     daoAddress: HexAddress
     pluginAddress: HexAddress
     network: NetworksEnum
-  }): Promise<Member> => {
+  }): Promise<Member | null> => {
+    const memberAddress = Web3Helper.parseAddress(params.memberAddress)
+    if (!memberAddress) return null
+
     const member = await ProxyMember.createMember(params.memberAddress)
 
-    const queryParams = {
-      memberAddress: member.address,
-      daoAddress: params.daoAddress,
-      pluginAddress: params.pluginAddress,
-      network: params.network,
-    }
-    const existingDao = await Models.DaoMemberMapping.findMapping(queryParams)
+    try {
+      return await DbTx.executeTxFn(async ({ session }) => {
+        const queryParams = {
+          memberAddress: params.memberAddress,
+          daoAddress: params.daoAddress,
+          pluginAddress: params.pluginAddress,
+          network: params.network,
+        }
+        const existingDao = await Models.DaoMemberMapping.findMapping(queryParams, { session })
 
-    if (existingDao) {
-      await DbTx.executeTxFn(async ({ session }) => {
-        const logDb = await existingDao.removeSelf({ session })
-        await session.commitTransaction()
-        logger.verbose('Remove DaoMemberMapping', llo({ logId: logDb.id }))
+        if (existingDao) {
+          const logDb = await existingDao.removeSelf({ session })
+          await session.commitTransaction()
+          await session.endSession()
+          logger.verbose('Remove DaoMemberMapping', llo({ logId: logDb.id }))
+        }
+
+        return member
       })
+    } catch (error) {
+      logger.error('Error in removeFromDao', llo({ error, params }))
+      return null
     }
-
-    return member
   },
 }
